Do not exit if commands return non-ExitCoder errors
Doing so limits the ability for users to have only some of their errors
cause the application to terminate while allowing others to bubble up.

This was originally an adjustment to #361 in #496 to fix #495, but, in
hindsight, I believe that the better approach is to recommend the use of
`RunAndExitOnError` for this use case (which is undeprecated here).

Fixes #565 #594